refactor: move home.tsx to pages folder
