+++ pages/index.tsx
@@ -1,7 +1,7 @@
 // pages/index.tsx
 import React from 'react';
 import { useSelector } from 'react-redux';
-import Home from '../src/components/Home';
+import Home from './home';
 import Header from '../src/components/Header/Header';
 import Flyout from '../src/components/Flyout/Flyout';
 import { RootState } from '../src/store/store';
